Reorganize code deleted unnecessary modules
