+++ lib/key.rb
@@ -1,13 +1,13 @@
-require './lib/key_generator'
-
 class Key
-
-include KeyGenerator
 
   attr_reader :key
 
   def initialize
     @key = generate_key
+  end
+
+  def generate_key
+    rand(10001).to_s.rjust(5, "0")
   end
 
   def a_key
