--- lib/key_generator.rb
@@ -1,10 +1,0 @@
-module KeyGenerator
-
-  def generate_key
-    rand(10001).to_s.rjust(5, "0")
-  end
-
-  def converted_key
-    generate_key.to_i
-  end
-end
