+++ lib/offset.rb
@@ -1,11 +1,13 @@
-require './lib/offset_generator'
-
 class Offset
-
-  include OffsetGenerator
 
   def initialize(date)
     @offset = generate_offset(date)
+  end
+
+  def generate_offset(date)
+    squared_date = date.delete("-").to_i**2
+    offset_string = squared_date.to_s[-4..-1]
+    offset = offset_string.to_i
   end
 
   def offset_a
